move sprite utils from encoding to common
