+++ modules/encoding/src/sprites/index.ts
@@ -1,3 +1,2 @@
 export * from "./sprite.parse";
 export * from "./sprite.stringify";
-export * from "./sprite.board2sprites";
